requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pyswisseph==2.10.3.2
pydantic==2.5.0